+++ SocialMediaFreelas.App/Application/InputModels/VagaInputModel.cs
@@ -4,7 +4,7 @@ public class VagaInputModel
     public string Descricao { get; set; }
     public string Cargo { get; set; }
     public string Tipo { get; set; }
-    public decimal Remuneracao { get; set; }
+    public double Remuneracao { get; set; }
     public int FreelancerId { get; set; }
     public int ClienteId { get; set; }
 }
