+++ SocialMediaFreelas.App/Application/UpdateModels/ExperienciaUpdateModel.cs
@@ -4,6 +4,6 @@ public class ExperienciaUpdateModel
     public string Projeto { get; set; }
     public string Empresa { get; set; }
     public string Tecnologia { get; set; }
-    public decimal Valor { get; set; }
+    public double Valor { get; set; }
     public int Avaliacao { get; set; }
 }
