+++ SocialMediaFreelas.App/Application/UpdateModels/FreelancerUpdateModel.cs
@@ -9,5 +9,5 @@ public class FreelancerUpdateModel : IDefaultEntity
     public DateTime DataNascimento { get; set; }
     public string Email { get; set; }
     public string Telefone { get; set; }
-    public decimal PretensaoSalarial { get; set; }
+    public double PretensaoSalarial { get; set; }
 }
