+++ SocialMediaFreelas.App/Core/Entities/Experiencia.cs
@@ -1,6 +1,6 @@
 public class Experiencia : BaseEntity
 {
-    public Experiencia(string projeto, string empresa, string tecnologia, decimal valor, int avaliacao, int freelancerId = 0)
+    public Experiencia(string projeto, string empresa, string tecnologia, double valor, int avaliacao, int freelancerId = 0)
     {
         FreelancerId = freelancerId;
         Projeto = projeto;
@@ -15,6 +15,6 @@ public class Experiencia : BaseEntity
     public string Projeto { get; private set; }
     public string Empresa { get; private set; }
     public string Tecnologia { get; private set; }
-    public decimal Valor { get; private set; }
+    public double Valor { get; private set; }
     public int Avaliacao { get; private set; }
 }
